backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
python-dotenv
passlib[bcrypt]==1.7.4
python-multipart